updateWebShopThoiTrang

+++ database/migrations/2022_07_08_023030_create_bill_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('users_id')->references('id')->on('users')->onDelete('cascade');
             $table->date('order');
             $table->string('address');
-            $table->decimal('totalmoney');
+            $table->integer('totalmoney');
             $table->timestamps();
         });
     }
